% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frailty.R
\docType{data}
\name{frailty}
\alias{frailty}
\title{Frailty in high dimensional survival data.}
\format{
A \code{tibble} with 7 columns and 272 rows which are :
\describe{
\item{institute}{Institute of the sample observations}
\item{del}{Numberic values 0 or 1 containing death/event information}
\item{timevar}{Survival duration}
\item{time.min}{Minimum survival}
\item{female}{Covariate_1, gender variable indicating either a female or not}
\item{ph.karno}{Covariate_2}
\item{pat.karno}{Covariate_3}}
}
\usage{
data(frailty)
}
\description{
Data set listing institutional wise survival outcomes

Survival observations data for frailty model functions of SurviMChd
}
\examples{
data(frailty)

}
\keyword{datasets}
