% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survweibMC.R
\name{survweibMC}
\alias{survweibMC}
\title{Weibull survival analysis with MCMC}
\usage{
survweibMC(m1, n1, m2, n2, chains, iter, data)
}
\arguments{
\item{m1}{Starting column number from where variables of high dimensional data will be selected.}

\item{n1}{Ending column number till where variables of high dimensional data will get selected.}

\item{m2}{Starting column number from where demographic observations starts}

\item{n2}{Ending column number of the demographic observations}

\item{chains}{Number of MCMC chains}

\item{iter}{Number of MCMC iterations}

\item{data}{High dimensional data having survival duration as (OS), event information as Death (1 if died, or 0 if alive).}
}
\value{
beta1[1] Posterior estimates of regression coefficients and deviance
}
\description{
Survival analysis with weibull distribution by MCMC
}
\examples{
\dontrun{
##
data(headnneck)
survweibMC(8,12,4,7,2,10,headnneck)
##
}
}
\references{
Kumar, M., Sonker, P. K., Saroj, A., Jain, A., Bhattacharjee, A.,
 & Saroj, R. K. (2020). Parametric survival analysis using R: Illustration with
 lung cancer data. \emph{Cancer Reports}, \bold{3(4)}, e1210.

Khan, S. A. (2018). Exponentiated Weibull regression for time-to-event
 data. \emph{Lifetime data analysis}, \bold{24(2)}, 328-354.
}
\seealso{
survexpMC
}
\author{
Atanu Bhattacharjee and Akash Pawar
}
