% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcsurv.R
\docType{data}
\name{mcsurv}
\alias{mcsurv}
\title{Metronomic cancer data}
\format{
A \code{tibble} with 15 columns which are :
\describe{
\item{OS}{Overall Survival}
\item{Death}{Death status}
\item{t}{Time at which event occurred}
\item{Covariate_1}{Variable measured on continuous scale}
\item{Covariate_2}{Variable measured on discrete scale}
\item{Covariate_3}{Variable measured on continuous scale}
\item{Covariate_4}{Variable measured on discrete scale}
\item{Covariate_5}{Variable measured on continuous scale}}
}
\usage{
data(mcsurv)
}
\description{
Observations made \code{tibble} on the head and neck cancer patients. Data for survMC function from SurviMChd package.
}
\examples{
data(mcsurv)
}
\keyword{datasets}
